Refactoring and add text 🔀
